Move TemplateHub to its own module
